+++ changelog.md
@@ -2,6 +2,7 @@
 - Replaced the Mario & Luigi costumes with a generic "male costume"
   - Dye & trim support
 - Christmas Hat now has dye & trim support
+- Koopa Shoes now has trim support
 
 ## Update 1.10.0
 - Added the Porcupuffer